+++ app/commands/decidim/challenges/survey_challenge.rb
@@ -3,7 +3,7 @@
 module Decidim
   module Challenges
     # This command is executed when the user answer a survey.
-    class SurveyChallenge < Rectify::Command
+    class SurveyChallenge < Decidim::Command
       # Initializes a SurveyChallenge Command.
       #
       # challenge - The current instance of the challenge to be answer the survey.
@@ -20,13 +20,18 @@ module Decidim
       #
       # Broadcasts :ok if successful, :invalid otherwise.
       def call
-        challenge.with_lock do
-          return broadcast(:invalid_form) unless survey_form.valid?
+        return broadcast(:invalid) unless can_answer_survey?
+        return broadcast(:invalid_form) unless survey_form.valid?
 
+        challenge.with_lock do
           answer_questionnaire
           create_survey
         end
+
         broadcast(:ok)
+      rescue ActiveRecord::Rollback
+        form.errors.add(:base, :invalid)
+        broadcast(:invalid)
       end
 
       private
@@ -36,7 +41,11 @@ module Decidim
       def answer_questionnaire
         return unless questionnaire?
 
-        Decidim::Forms::AnswerQuestionnaire.call(survey_form, user, challenge.questionnaire)
+        Decidim::Forms::AnswerQuestionnaire.call(survey_form, user, challenge.questionnaire) do
+          on(:invalid) do
+            raise ActiveRecord::Rollback
+          end
+        end
       end
 
       def create_survey
@@ -47,7 +56,11 @@ module Decidim
       end
 
       def questionnaire?
-        survey_form.model_name == "Questionnaire"
+        survey_form.model_name.human == "Questionnaire"
+      end
+
+      def can_answer_survey?
+        Decidim::Challenges::Survey.where(decidim_user_id: user, decidim_challenge_id: challenge).none?
       end
     end
   end