+++ app/packs/entrypoints/decidim_sdgs.js
@@ -1,0 +1,6 @@
+// Images
+require.context("../images", true)
+
+import "entrypoints/decidim_sdgs.scss";
+
+import "src/decidim/sdgs/filter.js";