+++ spec/system/decidim/problems/problems_global_search_spec.rb
@@ -6,12 +6,11 @@ describe "Problems global search", type: :system do
   include_context "with a component"
   let(:manifest_name) { "problems" }
   let!(:searchables) { create_list(:problem, 3, component: component) }
-  let!(:term) { translated(searchables.first.title).split(" ").last }
+  let!(:term) { translated(searchables.first.title).split.last }
 
   before do
     searchables.each { |s| s.update(published_at: Time.current) }
   end
 
-  # Failing tests
-  # include_examples "searchable results"
+  include_examples "unreportable searchable results"
 end